graph-rag-main/apps/agent_flow/requirements.txt
devtools==0.12.2
environs==11.0.0
fastapi==0.109.0
requests==2.31.0
uvicorn==0.27.0
tqdm==4.66.1

aiohttp==3.9.3
pydantic==2.6.1

redis==5.0.1

uule_grabber==0.1.9

pytest==8.2.0
pytest-asyncio==0.23.6

langchain[all]==0.2.10
langchain-openai==0.1.17
langchain-community==0.2.9
langchainhub==0.1.20
python-dotenv==1.0.1
aioredis==2.0.1
pytest-mock==3.14.0

scikit-learn==1.5.0
nltk==3.8.1
#watchtower==3.2.0
pandas==2.2.2